# docs/swagger.yaml
basePath: /api/v1
definitions:
  apiresponse.ErrorInfo:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  apiresponse.StandardResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/apiresponse.ErrorInfo'
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  v1.CreateEmployeeRequest:
    properties:
      hired_date:
        description: |-
          Date when the employee was hired (YYYY-MM-DD)
          example: 2024-01-15
        type: string
      name:
        description: |-
          Employee full name
          example: John Doe
        type: string
      position:
        description: |-
          Job position assigned to the employee
          example: Software Engineer
        type: string
      salary:
        description: |-
          Monthly salary of the employee
          example: 60000
        type: integer
    type: object
  v1.CreateEmployeeResponse:
    properties:
      created_at:
        description: 'example: 2024-01-15T10:30:00Z'
        type: string
      hired_date:
        description: 'example: 2024-01-15'
        type: string
      id:
        description: |-
          Employee ID generated by the system
          example: 1
        type: integer
      name:
        description: 'example: John Doe'
        type: string
      position:
        description: 'example: Software Engineer'
        type: string
      salary:
        description: 'example: 60000'
        type: integer
    type: object
  v1.CreateEmployeeResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/v1.CreateEmployeeResponse'
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  v1.GetAllEmployeesResponse:
    properties:
      created_at:
        description: 'example: 2024-01-15T10:30:00Z'
        type: string
      hired_date:
        description: 'example: 2024-01-15'
        type: string
      id:
        description: 'example: 1'
        type: integer
      name:
        description: 'example: John Doe'
        type: string
      position:
        description: 'example: Software Engineer'
        type: string
      salary:
        description: 'example: 60000'
        type: integer
    type: object
  v1.GetAllEmployeesResponseWrapper:
    properties:
      data:
        items:
          $ref: '#/definitions/v1.GetAllEmployeesResponse'
        type: array
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  v1.GetEmployeeByIdResponse:
    properties:
      created_at:
        description: 'example: 2024-01-15T10:30:00Z'
        type: string
      hired_date:
        description: 'example: 2024-01-15'
        type: string
      id:
        description: 'example: 1'
        type: integer
      name:
        description: 'example: John Doe'
        type: string
      position:
        description: 'example: Software Engineer'
        type: string
      salary:
        description: 'example: 60000'
        type: integer
    type: object
  v1.GetEmployeeByIdResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/v1.GetEmployeeByIdResponse'
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  v1.UpdateEmployeeRequest:
    properties:
      hired_date:
        description: 'example: 2024-01-15'
        type: string
      name:
        description: 'example: John Doe Updated'
        type: string
      position:
        description: 'example: Senior Software Engineer'
        type: string
      salary:
        description: 'example: 75000'
        type: integer
    type: object
  v1.UpdateEmployeeResponse:
    properties:
      hired_date:
        description: 'example: 2024-01-15'
        type: string
      id:
        description: 'example: 1'
        type: integer
      name:
        description: 'example: John Doe Updated'
        type: string
      position:
        description: 'example: Senior Software Engineer'
        type: string
      salary:
        description: 'example: 75000'
        type: integer
      updated_at:
        description: 'example: 2024-02-01T12:00:00Z'
        type: string
    type: object
  v1.UpdateEmployeeResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/v1.UpdateEmployeeResponse'
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for Employee Management
  title: Employee Management API
  version: "1.0"
paths:
  /employees:
    get:
      description: Retrieve a list of all employees in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GetAllEmployeesResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiresponse.StandardResponse'
      summary: Get all employees
      tags:
      - employees
    post:
      consumes:
      - application/json
      description: Creates a new employee and stores it in the database.
      parameters:
      - description: Employee create payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.CreateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.CreateEmployeeResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apiresponse.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiresponse.StandardResponse'
      summary: Create a new employee
      tags:
      - Employees
  /employees/{id}:
    delete:
      description: Remove an employee record by its ID
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apiresponse.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apiresponse.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiresponse.StandardResponse'
      summary: Delete an employee
      tags:
      - employees
    get:
      description: Fetch a single employee using the ID provided in the URL path
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GetEmployeeByIdResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apiresponse.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apiresponse.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiresponse.StandardResponse'
      summary: Get employee by ID
      tags:
      - Employees
    put:
      consumes:
      - application/json
      description: Update employee details by ID
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update employee payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UpdateEmployeeResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apiresponse.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apiresponse.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiresponse.StandardResponse'
      summary: Update an employee
      tags:
      - Employees
swagger: "2.0"