# internal/delivery/http/v1/dto.go
package v1

// CreateEmployeeRequest represents the payload for creating an employee.
// swagger:model CreateEmployeeRequest
type CreateEmployeeRequest struct {
	// Employee full name
	// example: John Doe
	Name string `json:"name"`

	// Job position assigned to the employee
	// example: Software Engineer
	Position string `json:"position"`

	// Monthly salary of the employee
	// example: 60000
	Salary int `json:"salary"`

	// Date when the employee was hired (YYYY-MM-DD)
	// example: 2024-01-15
	HiredDate string `json:"hired_date"`
}

// CreateEmployeeResponse represents the response returned after creating an employee.
// swagger:model CreateEmployeeResponse
type CreateEmployeeResponse struct {
	// Employee ID generated by the system
	// example: 1
	ID int `json:"id"`

	// example: John Doe
	Name string `json:"name"`

	// example: Software Engineer
	Position string `json:"position"`

	// example: 60000
	Salary int `json:"salary"`

	// example: 2024-01-15
	HiredDate string `json:"hired_date"`

	// example: 2024-01-15T10:30:00Z
	CreatedAt string `json:"created_at"`
}

// GetEmployeeByIdRequest represents the input for fetching an employee.
// swagger:model GetEmployeeByIdRequest
type GetEmployeeByIdRequest struct {
	// ID of the employee
	// example: 1
	ID int `json:"id"`
}

// GetEmployeeByIdResponse represents the returned employee record.
// swagger:model GetEmployeeByIdResponse
type GetEmployeeByIdResponse struct {
	// example: 1
	ID int `json:"id"`

	// example: John Doe
	Name string `json:"name"`

	// example: Software Engineer
	Position string `json:"position"`

	// example: 60000
	Salary int `json:"salary"`

	// example: 2024-01-15
	HiredDate string `json:"hired_date"`

	// example: 2024-01-15T10:30:00Z
	CreatedAt string `json:"created_at"`
}

// GetAllEmployeesResponse represents a single employee entry in a list.
// swagger:model GetAllEmployeesResponse
type GetAllEmployeesResponse struct {
	// example: 1
	ID int `json:"id"`

	// example: John Doe
	Name string `json:"name"`

	// example: Software Engineer
	Position string `json:"position"`

	// example: 60000
	Salary int `json:"salary"`

	// example: 2024-01-15
	HiredDate string `json:"hired_date"`

	// example: 2024-01-15T10:30:00Z
	CreatedAt string `json:"created_at"`
}

// UpdateEmployeeRequest contains fields for updating an employee.
// swagger:model UpdateEmployeeRequest
type UpdateEmployeeRequest struct {
	// example: John Doe Updated
	Name string `json:"name"`

	// example: Senior Software Engineer
	Position string `json:"position"`

	// example: 75000
	Salary int `json:"salary"`

	// example: 2024-01-15
	HiredDate string `json:"hired_date"`
}

// UpdateEmployeeResponse returned after updating an employee.
// swagger:model UpdateEmployeeResponse
type UpdateEmployeeResponse struct {
	// example: 1
	ID int `json:"id"`

	// example: John Doe Updated
	Name string `json:"name"`

	// example: Senior Software Engineer
	Position string `json:"position"`

	// example: 75000
	Salary int `json:"salary"`

	// example: 2024-01-15
	HiredDate string `json:"hired_date"`

	// example: 2024-02-01T12:00:00Z
	UpdatedAt string `json:"updated_at"`
}